refactor: extract MeshBufferUploadHelper to eliminate duplicated EnsureMeshBuffers
