+++ src/Beutl.Engine/Graphics3D/Nodes/GeometryPass.cs
@@ -1,10 +1,8 @@
 ﻿using System.Numerics;
-using System.Runtime.InteropServices;
 using Beutl.Composition;
 using Beutl.Graphics.Backend;
 using Beutl.Graphics3D.Lighting;
 using Beutl.Graphics3D.Materials;
-using Beutl.Graphics3D.Meshes;
 using Beutl.Media;
 
 namespace Beutl.Graphics3D.Nodes;
@@ -170,7 +168,7 @@ public sealed class GeometryPass : GraphicsNode3D
             return;
 
         // Ensure GPU buffers are created/updated
-        EnsureMeshBuffers(meshResource);
+        MeshBufferUploadHelper.Ensure(Context, meshResource);
 
         if (meshResource.VertexBuffer == null || meshResource.IndexBuffer == null)
             return;
@@ -192,59 +190,6 @@ public sealed class GeometryPass : GraphicsNode3D
 
         // Draw the mesh
         RenderPass.DrawIndexed((uint)meshResource.IndexCount);
-    }
-
-    private void EnsureMeshBuffers(Mesh.Resource meshResource)
-    {
-        if (!meshResource.BuffersDirty)
-            return;
-
-        var vertices = meshResource.GetVertices();
-        var indices = meshResource.GetIndices();
-
-        if (vertices.Length == 0 || indices.Length == 0)
-            return;
-
-        ulong vertexSize = (ulong)(vertices.Length * Marshal.SizeOf<Vertex3D>());
-        ulong indexSize = (ulong)(indices.Length * sizeof(uint));
-
-        // Dispose old buffers if they exist
-        meshResource.VertexBuffer?.Dispose();
-        meshResource.IndexBuffer?.Dispose();
-
-        // Create new device-local buffers
-        var vertexBuffer = Context.CreateBuffer(
-            vertexSize,
-            BufferUsage.VertexBuffer | BufferUsage.TransferDestination,
-            MemoryProperty.DeviceLocal);
-
-        var indexBuffer = Context.CreateBuffer(
-            indexSize,
-            BufferUsage.IndexBuffer | BufferUsage.TransferDestination,
-            MemoryProperty.DeviceLocal);
-
-        // Create staging buffers and upload
-        using var vertexStaging = Context.CreateBuffer(
-            vertexSize,
-            BufferUsage.TransferSource,
-            MemoryProperty.HostVisible | MemoryProperty.HostCoherent);
-
-        using var indexStaging = Context.CreateBuffer(
-            indexSize,
-            BufferUsage.TransferSource,
-            MemoryProperty.HostVisible | MemoryProperty.HostCoherent);
-
-        vertexStaging.Upload(vertices);
-        indexStaging.Upload(indices);
-
-        // Copy from staging to device local
-        Context.CopyBuffer(vertexStaging, vertexBuffer, vertexSize);
-        Context.CopyBuffer(indexStaging, indexBuffer, indexSize);
-
-        // Store in mesh resource
-        meshResource.VertexBuffer = vertexBuffer;
-        meshResource.IndexBuffer = indexBuffer;
-        meshResource.BuffersDirty = false;
     }
 
     protected override void OnDispose()
