+++ src/Beutl.Engine/Graphics3D/Nodes/MeshBufferUploadHelper.cs
@@ -1,0 +1,56 @@
+﻿using System.Runtime.InteropServices;
+using Beutl.Graphics.Backend;
+using Beutl.Graphics3D.Meshes;
+
+namespace Beutl.Graphics3D.Nodes;
+
+internal static class MeshBufferUploadHelper
+{
+    public static void Ensure(IGraphicsContext context, Mesh.Resource meshResource)
+    {
+        if (!meshResource.BuffersDirty)
+            return;
+
+        var vertices = meshResource.GetVertices();
+        var indices = meshResource.GetIndices();
+
+        if (vertices.Length == 0 || indices.Length == 0)
+            return;
+
+        ulong vertexSize = (ulong)(vertices.Length * Marshal.SizeOf<Vertex3D>());
+        ulong indexSize = (ulong)(indices.Length * sizeof(uint));
+
+        meshResource.VertexBuffer?.Dispose();
+        meshResource.IndexBuffer?.Dispose();
+
+        var vertexBuffer = context.CreateBuffer(
+            vertexSize,
+            BufferUsage.VertexBuffer | BufferUsage.TransferDestination,
+            MemoryProperty.DeviceLocal);
+
+        var indexBuffer = context.CreateBuffer(
+            indexSize,
+            BufferUsage.IndexBuffer | BufferUsage.TransferDestination,
+            MemoryProperty.DeviceLocal);
+
+        using var vertexStaging = context.CreateBuffer(
+            vertexSize,
+            BufferUsage.TransferSource,
+            MemoryProperty.HostVisible | MemoryProperty.HostCoherent);
+
+        using var indexStaging = context.CreateBuffer(
+            indexSize,
+            BufferUsage.TransferSource,
+            MemoryProperty.HostVisible | MemoryProperty.HostCoherent);
+
+        vertexStaging.Upload(vertices);
+        indexStaging.Upload(indices);
+
+        context.CopyBuffer(vertexStaging, vertexBuffer, vertexSize);
+        context.CopyBuffer(indexStaging, indexBuffer, indexSize);
+
+        meshResource.VertexBuffer = vertexBuffer;
+        meshResource.IndexBuffer = indexBuffer;
+        meshResource.BuffersDirty = false;
+    }
+}
