+++ src/Beutl.Engine/Graphics3D/Nodes/PointShadowPass.cs
@@ -374,7 +374,7 @@ void main() {
         if (meshResource == null)
             return;
 
-        EnsureMeshBuffers(meshResource);
+        MeshBufferUploadHelper.Ensure(Context, meshResource);
 
         if (meshResource.VertexBuffer == null || meshResource.IndexBuffer == null)
             return;
@@ -393,59 +393,6 @@ void main() {
 
         // Draw the mesh
         RenderPass.DrawIndexed((uint)meshResource.IndexCount);
-    }
-
-    private void EnsureMeshBuffers(Mesh.Resource meshResource)
-    {
-        if (!meshResource.BuffersDirty)
-            return;
-
-        var vertices = meshResource.GetVertices();
-        var indices = meshResource.GetIndices();
-
-        if (vertices.Length == 0 || indices.Length == 0)
-            return;
-
-        ulong vertexSize = (ulong)(vertices.Length * Marshal.SizeOf<Vertex3D>());
-        ulong indexSize = (ulong)(indices.Length * sizeof(uint));
-
-        // Dispose old buffers if they exist
-        meshResource.VertexBuffer?.Dispose();
-        meshResource.IndexBuffer?.Dispose();
-
-        // Create new device-local buffers
-        var vertexBuffer = Context.CreateBuffer(
-            vertexSize,
-            BufferUsage.VertexBuffer | BufferUsage.TransferDestination,
-            MemoryProperty.DeviceLocal);
-
-        var indexBuffer = Context.CreateBuffer(
-            indexSize,
-            BufferUsage.IndexBuffer | BufferUsage.TransferDestination,
-            MemoryProperty.DeviceLocal);
-
-        // Create staging buffers and upload
-        using var vertexStaging = Context.CreateBuffer(
-            vertexSize,
-            BufferUsage.TransferSource,
-            MemoryProperty.HostVisible | MemoryProperty.HostCoherent);
-
-        using var indexStaging = Context.CreateBuffer(
-            indexSize,
-            BufferUsage.TransferSource,
-            MemoryProperty.HostVisible | MemoryProperty.HostCoherent);
-
-        vertexStaging.Upload(vertices);
-        indexStaging.Upload(indices);
-
-        // Copy from staging to device local
-        Context.CopyBuffer(vertexStaging, vertexBuffer, vertexSize);
-        Context.CopyBuffer(indexStaging, indexBuffer, indexSize);
-
-        // Store in mesh resource
-        meshResource.VertexBuffer = vertexBuffer;
-        meshResource.IndexBuffer = indexBuffer;
-        meshResource.BuffersDirty = false;
     }
 
     protected override void OnDispose()
